Add work 5 task 3

+++ Programming С++/Work_5/sqrt_vector/sqrt_vector.cpp
@@ -45,7 +45,7 @@ int main()
 	int i{ 4 };
 	auto res{ Func(i) };
 	cout << "OUT:\t" << res << endl;
-	vector<int> vec{ -1, 4, 8 };
+	vector<double> vec{ -1, 4, 8 };
 	cout << "IN:\t";
 	for (const auto v : vec) { cout << v << " "; };
 	cout << endl;
